+++ src/core/models/comment.ts
@@ -3,4 +3,5 @@ import { CommentRange } from 'typescript';
 export interface Comment {
     text: string;
     range: CommentRange;
+    isTripleSlashDirective: boolean;
 }