+++ src/core/models/parsed-import-values.ts
@@ -1,0 +1,6 @@
+import { ImportElement } from './import-element';
+
+export interface ParsedImportValues {
+    importElements: ImportElement[];
+    usedTypeReferences: string[];
+}